create DragAndDropPage, its elements, and methods
