+++ src/test/java/de/frameworktsr/pages/ApiDemosPage.java
@@ -7,26 +7,17 @@ import io.appium.java_client.AppiumBy;
 public class ApiDemosPage extends BasePage {
     By viewsLinkBy = AppiumBy.accessibilityId("Views");
     By dragAndDropLinkBy = AppiumBy.accessibilityId("Drag and Drop");
-    By textlessCircleBy = AppiumBy.id("io.appium.android.apis:id/drag_dot_1");
-    By dragAndDropMessageContainerBy = AppiumBy.id("io.appium.android.apis:id/drag_result_text");
+    
 
     public void navigateToviews() {
         click(viewsLinkBy);
     }
 
-    public void navigateToDragAndDropPage() {
+    public DragAndDropPage navigateToDragAndDropPage() {
         click(dragAndDropLinkBy);
+
+        return new DragAndDropPage();
     }
 
-    public void dragAndDropTextLessCircle() {
-        dragAndDropUsingElementId(textlessCircleBy, 800, 1300);
-        // dragAndDropUsingCoordinates(271, 790, 1000, 2507);
-    }
-
-    public Boolean isDragAndDropSuccessful(String expectedMessage) {
-        String actualMessage = getTextAttribute(dragAndDropMessageContainerBy);
-        checkSimilarityOfText(expectedMessage, actualMessage);
-
-        return true;
-    }
+   
 }
