+++ src/test/java/de/frameworktsr/stepdefinitions/DragAndDrop.java
@@ -1,11 +1,13 @@
 package de.frameworktsr.stepdefinitions;
 
 import de.frameworktsr.pages.ApiDemosPage;
+import de.frameworktsr.pages.DragAndDropPage;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 
 public class DragAndDrop {
     ApiDemosPage apiDemosPage = new ApiDemosPage();
+    DragAndDropPage dragDropPage;
 
     @When("the user navigates to Views")
     public void userNavigatesToViews() {
@@ -14,16 +16,16 @@ public class DragAndDrop {
 
     @When("the user naviagates to Drag and Drop page")
     public void userNavigatesToDragAndDropPage() {
-        apiDemosPage.navigateToDragAndDropPage();
+        dragDropPage = apiDemosPage.navigateToDragAndDropPage();
     }
 
     @When("the user drags and drops the textless circle")
     public void userDragsAndDropTextlessCircle() {
-        apiDemosPage.dragAndDropTextLessCircle();
+        dragDropPage.dragAndDropTextLessCircle();
     }
 
     @Then("the message {string} should be displayed")
     public void theMessageShouldBeDisplayed(String string) {
-        apiDemosPage.isDragAndDropSuccessful(string);
+        dragDropPage.isDragAndDropSuccessful(string);
     }
 }
